Vastly Simplified GridTimer
Removed printing function from the class. Instead added a logger
delegate, which will output all necessary timing info to the logger, if
there is one.
I'm not completely done with GridTimer, I need to test the heatmap
functionality, and I also intend to add a speedup calculator, which
checks the average time per pixel and compares that to overall ray
casting time.