First unit tests
This is hardly critical for now, but it does feel like it's worth my
time to write unit tests to test functionality rather than the little
temporary tests inside the Tracer class.
I would like to finish the implementation of the ray tracer before
fleshing this out fully, but in the meantime I may choose to test new
features in this branch.