Updates to Vector, Ray, Color and GridTimer
In GridTimer I reduced the precision of the output.
For Vector I changed add() to sum() with variable number of arguments.
For Ray I added a toString() for use while debugging.
for Color, I added added a few pieces of functionality, fixed mix() to
perform the correct action, and added a nice toString() for debugging.